fix(api): nit leftover
Ref. metriport/metriport-internal#1785
