+++ packages/api/src/external/tsconfig.json
@@ -1,8 +1,8 @@
 {
   "extends": "../../tsconfig.json",
   "compilerOptions": {
-    "noUncheckedIndexedAccess": true
+    "noUncheckedIndexedAccess": true,
   },
-  "include": ["./**/*", "../command/medical/facility/create-or-update-facility.ts"],
-  "exclude": ["./**/*.test.ts"]
+  "include": ["./**/*"],
+  "exclude": ["./**/*.test.ts"],
 }
